Move Doctor details page
