+++ src/App.js
@@ -2,7 +2,7 @@ import './styles/App.scss';
 import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import MainPage from './components/HomePage/MainPage';
-import Details from './components/DetailsPage/Details';
+import Details from './components/DoctorsData/DetailsPage/Details';
 import Reservation from './components/ReservationPage/ReservationPage';
 import LoginPage from './components/Authentication/LoginPage';
 import RegistrionPage from './components/Authentication/RegistrionPage';
