+++ src/components/DoctorsData/DetailsPage/Details.js
@@ -1,5 +1,5 @@
 import { useLocation } from 'react-router-dom';
-import NavBar from '../NavBar/Navbar';
+import NavBar from '../../NavBar/Navbar';
 
 function Details() {
   const location = useLocation();
